Add attachment#is_valid? test and implementation
attachment#is_valid? when attachment type is image/png it should be true.

attachment#is_valid? when attachment type is image/gif it should be true

attachment#is_valid? when attachment type is image/jpeg it should be true.

attachment#is_valid? when attachment type is text/plain it should be true.

attachment#is_valid? when attachment type is text/csv it should be true.

attachment#is_valid? when attachment type is application/x-tar it should be false.

Signed-off-by: Yudo Kartiko <yudoworks@gmail.com>